Python client Package restructure
Restructures the Python Traffic Control client into a real package - without links
. Also includes real documentation linking, real contact info for the package, and
adds missing endpoint references/fixes old/deprecated/broken ones, removed a couple
of duplicated endpoint methods and fixes some poorly-formatted reference labels in
the API documentation.
