+++ docs/source/api/cachegroups_id_parameters.rst
@@ -13,7 +13,7 @@
 .. limitations under the License.
 ..
 
-.. _to-api-cachegroup_id_parameters:
+.. _to-api-cachegroups-id-parameters:
 
 *********************************
 ``cachegroups/{{ID}}/parameters``
