+++ docs/source/conf.py
@@ -29,8 +29,11 @@ import os
 import sys
 here = os.path.dirname(__file__)
 ortPath = os.path.join(here, '..', '..', 'infrastructure', 'cdn-in-a-box', 'ort')
+clientPath = os.path.join(here, '..', '..', 'traffic_control', 'clients', 'python')
 ortPath = os.path.abspath(ortPath)
+clientPath = os.path.abspath(clientPath)
 sys.path.insert(0, ortPath)
+sys.path.insert(0, clientPath)
 
 # -- Custom Lexical Analyzer for DNS -----------------------------------------
