+++ traffic_control/clients/python/.gitignore
@@ -1,0 +1,3 @@
+*.egg-info
+build/
+dist/
