--- traffic_control/clients/python/trafficops/common
@@ -1,1 +1,0 @@
-../common/
