--- traffic_control/clients/python/trafficops/trafficops/__init__.py
@@ -1,27 +1,0 @@
-# -*- coding: utf-8 -*-
-
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# Core Modules
-import logging
-
-logger = logging.getLogger(__name__)
-
-# Import Local Modules into the 'trafficops' package namespace for convenience.
-from common.restapi import *
-from trafficops.tosession import *
-
-# The Module Version
-__version__ = '0.1.1'
